requirements.txt
fastapi
tensorflow
uvicorn
python-multipart
pillow
tensorflow-serving-api
matplotlib
numpy